Looping to generate lots of scopes example
I've been considering whether Consistent should do more to help with these patterns and I think the answer is no. The moment it gets into creating DOM for you it has stepped outside of its remit. There are lots of ways to create DOM, and one can always add utility functions on top of Consistent to support the patterns that you need.

Also trying out C as an abbreviation for Consistent.