Watchers are now called before the apply touches the DOM.
We now prevent re-entry into apply, and we loop multiple times across watchers in case they themselves change the scope. There is the potential for an infinite loop, which we detect (solving the halting problem) and throw an exception for.

We no longer actually apply to the DOM if there are no changes.

scope.$.update() no longer invokes watchers. You have to call apply.