jQuery plugin no longer automatically applies the scope.
This adds an extra step when you're using an existing scope. Maybe there should be an option to automatically apply if you've supplied a scope, perhaps an additional boolean. But it is probably best as not the default as it doesn't just apply to the new nodes, it applies the whole scope, so it can hide other problems - it did in the dom-as-source-multiple.html where because we were adding more nodes to the scope we implicitly applied which covered up that we were missing an apply.