scope.$.get and scope.$.set behaviour changes to address bug reported by @snikch relating to child scopes and value functions
Change scope.$.get method to evaluate value functions, in this way it behaves more like accessing a snapshot. This actually matches the way that it has been used; in watchers and initialising repeats.

Change scope.$.set so that it looks for value functions in parents to call to "set" the value. This is important, as we will try to set the value in the scope during an update, which occurs if you toggle a form element, and if that value was driven by a value function in a parent, then we would set a local static value and then that value would shadow the value function in the parent.