Mark nodes as dirty after an update event so we reapply the scope to the DOM
This is necessary if we have form elements bound to value functions; the user changes the form element, we call the value function and attempt to set it, it still returns the old value, we need to update the DOM element back, but because the scope hasn't changed we don't. Until now.