Controllers: event handler functions are always called with this set to the controller
Because of the way in which we find controller functions, we don't know the controller they are defined within. So it is hard to know what to set "this" to. I have chosen to wrap controller functions returned from scope.$.controller(name) in an anonymous function that forces "this" to be the controller.